ai_model/requirements.txt
pandas
numpy
scikit-learn
flask
notebook
requests
python-dotenv